requirements.txt
numpy==1.15.2
scipy==1.3.3
torch==1.2.0
torchvision==0.4.0
matplotlib==3.0.0
opencv_python==4.1.2.30
Cython==0.29.16
easydict==1.9
ipython==7.13.0
Pillow==7.1.1
